fix(ci): build linux/macos specific test not on Windows
